* vt_term.c: Missing '}' is fixed. * vt_parser.[ch]: The return value of vt_parser_write() is changed from   int to size_t. * ui_display_linux.c: Compiling error in no KT_DEAD2 macro definition is fixed. * c_sixel.c: SIXEL_ORMODE is defined.

+++ vtemu/vt_parser.c
@@ -7151,7 +7151,7 @@ int vt_parse_vt100_sequence(vt_parser_t *vt_parser) {
   return 1;
 }
 
-int vt_parser_write(vt_parser_t *vt_parser, u_char *buf, size_t len) {
+size_t vt_parser_write(vt_parser_t *vt_parser, u_char *buf, size_t len) {
   if (vt_parser->is_transferring_data) {
     return 0;
   }
